Added Materials and Formulas
I have added materials for Elastic and Inelastic collision which is added to the Cars on spawn, currently just to car two. A new car is spawned as well but that is an issue in the code right now that I plan to fix. I also added in formulas to the physics calculator.